+++ Runtime/Implementation/AccelByteJuiceTasks.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2023 - 2024 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2023 - 2025 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
@@ -17,9 +17,10 @@ namespace AccelByte.Networking
             state = inState;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceStateChanged(state);
+            logger?.Log($"{GetType().Name} state: {state}");
         }
     }
 
@@ -36,9 +37,10 @@ namespace AccelByte.Networking
             isSuccess = inIsSuccess;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceCandidateFound(sdp, isSuccess);
+            logger?.Log($"{GetType().Name} sdp: {sdp} - {isSuccess}");
         }
     }
 
@@ -51,9 +53,10 @@ namespace AccelByte.Networking
             juiceInstance = inJuiceInstance;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceGatheringDone();
+            logger?.Log($"{GetType().Name}");
         }
     }
 
@@ -70,9 +73,10 @@ namespace AccelByte.Networking
             size = inSize;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceDataReceived(data, size);
+            logger?.Log($"{GetType().Name} size {size}");
         }
     }
 
@@ -87,9 +91,11 @@ namespace AccelByte.Networking
             message = inMessage;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
-            AccelByteDebug.LogVerbose($"JUICE_NATIVE_LOG {logLevel}: {message}");
+            var logMessage = $"JUICE_NATIVE_LOG {logLevel}: {message}";
+            AccelByteDebug.LogVerbose(logMessage);
+            logger?.Log(logMessage);
         }
     }
 }