+++ Runtime/Implementation/AccelByteRTCPeerTask.cs
@@ -1,8 +1,9 @@
-// Copyright (c) 2024 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2025 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
 using System;
+using System.Drawing;
 using AccelByte.Core;
 namespace AccelByte.Networking
 {
@@ -17,9 +18,10 @@ namespace AccelByte.Networking
             state = inState;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceStateChanged(state);
+            logger?.Log($"{GetType().Name} state: {state}");
         }
     }
 
@@ -36,9 +38,10 @@ namespace AccelByte.Networking
             isSuccess = inIsSuccess;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceCandidateFound(sdp, isSuccess);
+            logger?.Log($"{GetType().Name} sdp: {sdp} - {isSuccess}");
         }
     }
 
@@ -51,9 +54,10 @@ namespace AccelByte.Networking
             juiceInstance = inJuiceInstance;
         }
 
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceGatheringDone();
+            logger?.Log($"{GetType().Name}");
         }
     }
 
@@ -69,10 +73,11 @@ namespace AccelByte.Networking
             data = inData;
             size = inSize;
         }
-    
-        public void Execute()
+
+        public void Execute(IDebugger logger)
         {
             juiceInstance?.OnJuiceDataReceived(data, size);
+            logger?.Log($"{GetType().Name} size {size}");
         }
     }
 
@@ -87,9 +92,11 @@ namespace AccelByte.Networking
             message = inMessage;
         }
     
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
-            AccelByteDebug.LogVerbose($"RTC_PEER_LOG {logLevel}: {message}");
+            var logMessage = $"RTC_PEER_LOG {logLevel}: {message}";
+            AccelByteDebug.LogVerbose(logMessage);
+            logger?.Log(logMessage);
         }
     }
 
@@ -108,9 +115,10 @@ namespace AccelByte.Networking
             localDescription = inDescription;
         }
         
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             callback?.Invoke(localDescription);
+            logger?.Log($"{GetType().Name} sdp {localDescription}");
         }
     }
     
@@ -129,9 +137,10 @@ namespace AccelByte.Networking
             setupResult = inResult;
         }
         
-        public void Execute()
+        public void Execute(IDebugger logger)
         {
             callback?.Invoke(setupResult);
+            logger?.Log($"{GetType().Name} result {setupResult}");
         }
     }
 }