+++ Runtime/Implementation/JuiceTask.cs
@@ -2,10 +2,12 @@
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
+using AccelByte.Core;
+
 namespace AccelByte.Networking
 {
     public interface JuiceTask
     {
-        public void Execute();
+        public void Execute(IDebugger logger);
     }
 }