+++ Runtime/Models/TestUtilsModels.cs
@@ -1,0 +1,13 @@
+// Copyright (c) 2025 AccelByte Inc. All Rights Reserved.
+// This is licensed software from AccelByte Inc, for limitations
+// and restrictions contact your company contract manager.
+
+using UnityEngine.Scripting;
+
+namespace AccelByte.Models
+{
+    [Preserve]
+    internal class P2POptionalParameters : OptionalParametersBase
+    {
+    }
+}